refactor(worker): extract pool turn-driver adapters + IO shim (A1)
Move _TurnDriver/_GraphTurnDriver to primer/worker/drivers.py and
_WorkspaceIOShim to primer/worker/io_shim.py (pure adapter classes, no
pool reference). Re-export all three from primer.worker.pool so existing
importers (executor_builders, tests) keep resolving the old paths. Pure
code motion, no behaviour change.
