+++ primer/worker/pool.py
@@ -32,6 +32,8 @@ from primer.model.scheduler import WorkerConfig
 from primer.model.workspace_session import WorkspaceSession, SessionStatus
 from primer.model.yield_ import YieldToWorker
 from primer.worker.turn import _CancelScope
+from primer.worker.drivers import _GraphTurnDriver, _TurnDriver  # noqa: F401  re-export
+from primer.worker.io_shim import _WorkspaceIOShim
 from primer.worker.yield_resume_registry import get_resume_hook
 from primer.worker.yield_runtime import (
     _resume_tool_approval,
@@ -1393,216 +1395,6 @@ class WorkerPool:
 
     def _infer_post_turn_status(self, executor, session: WorkspaceSession) -> SessionStatus:
         return executor_builders.infer_post_turn_status(self, executor, session)
-
-
-class _TurnDriver:
-    """Adapter that consumes a streaming executor for the turn-based pool.
-
-    The workspace executors expose ``invoke`` as an async generator
-    (yielding :class:`StreamEvent`s), but the pool dispatches turns via
-    ``await executor.invoke([])`` -- a single-shot coroutine. This adapter
-    bridges the two: drain the generator to completion inside an awaitable
-    ``invoke`` and surface :attr:`last_done_reason` from the underlying
-    executor for the post-turn status mapper to read.
-    """
-
-    def __init__(self, executor) -> None:
-        self._executor = executor
-
-    @property
-    def last_done_reason(self) -> str | None:
-        return getattr(self._executor, "last_done_reason", None)
-
-    @property
-    def session(self):
-        # Pass-through to support tests / introspection that want the
-        # underlying :class:`AgentSession`.
-        return getattr(self._executor, "session", None)
-
-    async def invoke(self, messages, *, response_format=None) -> None:
-        """Drain the executor's stream to completion.
-
-        Events are intentionally discarded here -- streaming-tap
-        subscribers attached via :meth:`_BaseAgentExecutor.subscribe`
-        still receive them via the executor's own fan-out.
-        """
-        async for _ev in self._executor.invoke(
-            messages, response_format=response_format
-        ):
-            pass
-
-
-class _GraphTurnDriver:
-    """Adapter for :class:`primer.graph.workspace_executor.WorkspaceGraphExecutor`.
-
-    Two differences from :class:`_TurnDriver`:
-
-    * Graph executor's ``invoke(messages)`` does NOT accept a
-      ``response_format`` kwarg (per-node response_format lives on
-      each agent node), so this adapter discards the kwarg the worker
-      passes uniformly.
-    * The graph executor runs the WHOLE graph in one ``invoke()`` call
-      (multiple supersteps complete internally before returning), so
-      ``last_done_reason`` is a fixed ``"graph_ended"`` sentinel that
-      :meth:`WorkerPool._infer_post_turn_status` recognises as ENDED
-      — the session is never re-enqueued.
-    """
-
-    def __init__(self, executor) -> None:
-        self._executor = executor
-
-    @property
-    def last_done_reason(self) -> str:
-        return "graph_ended"
-
-    @property
-    def session(self):
-        # Graphs have no single AgentSession in Phase 1; expose None
-        # so the pool's introspection callers see a consistent shape.
-        return getattr(self._executor, "_workspace_session", None)
-
-    async def invoke(self, messages, *, response_format=None) -> None:
-        """Drain the graph executor's stream to completion.
-
-        ``response_format`` is accepted for signature compatibility
-        with :class:`_TurnDriver` and silently discarded — graph nodes
-        carry their own per-node ``response_format`` on the
-        :class:`_AgentNodeRef` model.
-        """
-        async for _ev in self._executor.invoke(messages):
-            pass
-
-
-class _WorkspaceIOShim:
-    """``WorkspaceIO`` adapter that delegates to the workspace runtime.
-
-    Satisfies the :class:`primer.session.persistence.WorkspaceIO` protocol
-    used by :class:`WorkspaceMessageWriter`.
-
-    Dispatch: resolves the workspace from the registry and calls
-    ``workspace.append_message_line(session_id, line)`` directly.  Every
-    concrete :class:`primer.int.workspace.Workspace` backend now implements
-    this method (added in Task 9).
-
-    The shim calls ``_workspace_registry.get_workspace(workspace_id)`` on
-    each flush so hot-reloaded workspace instances are picked up during
-    long-lived workers.  Because the session_id alone is enough to identify
-    the session slot inside the workspace, but the workspace_id is needed to
-    locate the workspace, the shim tracks the mapping via a lightweight
-    ``_session_to_workspace`` dict populated via :meth:`register_session`
-    before the first write.
-    """
-
-    def __init__(self, workspace_registry) -> None:
-        self._registry = workspace_registry
-        # session_id -> workspace_id mapping; populated via register_session()
-        # from the _build_session_executor path before any append is called.
-        self._session_to_workspace: dict[str, str] = {}
-
-    def register_session(self, session_id: str, workspace_id: str) -> None:
-        """Pre-register the workspace_id for a session (called by the pool)."""
-        self._session_to_workspace[session_id] = workspace_id
-
-    async def append_message_line(self, session_id: str, line: bytes) -> None:
-        """Append ``line`` to the session's ``messages.jsonl`` via the workspace runtime."""
-        if self._registry is None:
-            logger.warning(
-                "_WorkspaceIOShim: no workspace_registry configured; "
-                "dropping %d bytes for session %s",
-                len(line), session_id,
-            )
-            return
-
-        workspace_id = self._session_to_workspace.get(session_id)
-        if workspace_id is None:
-            logger.warning(
-                "_WorkspaceIOShim: no workspace_id registered for session %s; "
-                "dropping %d bytes",
-                session_id, len(line),
-            )
-            return
-
-        workspace = await self._registry.get_workspace(workspace_id)
-        if workspace is None:
-            logger.warning(
-                "_WorkspaceIOShim: workspace %r not found for session %s; "
-                "dropping %d bytes",
-                workspace_id, session_id, len(line),
-            )
-            return
-
-        await workspace.append_message_line(session_id, line)
-
-    def workspace_id_for(self, session_id: str) -> str | None:
-        """Public lookup for the workspace id bound to a session.
-
-        Replaces direct reads of the private ``_session_to_workspace``
-        dict by call sites that need to resolve a session's workspace
-        (e.g. dispatch's turn-log factory closure).
-        """
-        return self._session_to_workspace.get(session_id)
-
-    async def append_state_line(
-        self, workspace_id: str, state_relative_path: str, line: bytes,
-    ) -> None:
-        """Append ``line`` to ``<workspace.state_path>/<state_relative_path>``.
-
-        Resolves the workspace via the registry, prepends the workspace's
-        own ``state_path`` (so operators can override the default
-        ``.state`` via :class:`WorkspaceTemplate` without losing the
-        writer/reader path agreement), then delegates to the backend's
-        ``append_state_line``. Logs and drops the bytes if the registry
-        is absent or the workspace can't be resolved (mirroring
-        ``append_message_line``'s best-effort policy).
-        """
-        if self._registry is None:
-            logger.warning(
-                "_WorkspaceIOShim: no workspace_registry configured; "
-                "dropping %d state bytes for workspace %s",
-                len(line), workspace_id,
-            )
-            return
-        workspace = await self._registry.get_workspace(workspace_id)
-        if workspace is None:
-            logger.warning(
-                "_WorkspaceIOShim: workspace %r not found; "
-                "dropping %d state bytes",
-                workspace_id, len(line),
-            )
-            return
-        state_path = getattr(workspace, "state_path", ".state")
-        full_path = f"{state_path}/{state_relative_path}"
-        try:
-            await workspace.append_state_line(full_path, line)
-        except NotImplementedError:
-            # Backend without turn-log support; silently no-op so
-            # the dispatch doesn't bubble the failure.
-            logger.debug(
-                "_WorkspaceIOShim: workspace %r has no append_state_line; "
-                "dropping %d bytes", workspace_id, len(line),
-            )
-
-    async def read_state_file(
-        self, workspace_id: str, state_relative_path: str,
-    ) -> bytes:
-        """Read ``<workspace.state_path>/<state_relative_path>`` from the workspace.
-
-        Returns ``b""`` when the workspace is gone, the path doesn't
-        exist, or any other backend error fires. Used by the turn-log
-        writer's lazy bootstrap so the same path-resolution rule
-        applies to both reads and writes.
-        """
-        if self._registry is None:
-            return b""
-        workspace = await self._registry.get_workspace(workspace_id)
-        if workspace is None:
-            return b""
-        state_path = getattr(workspace, "state_path", ".state")
-        full_path = f"{state_path}/{state_relative_path}"
-        try:
-            return await workspace.read_file(full_path)
-        except Exception:  # noqa: BLE001
-            return b""
 
 
 # Imported at the bottom so the helper modules (which import names defined
